新增关注功能

+++ model/userFollow.go
@@ -12,8 +12,8 @@ package model
 import "time"
 
 type UserFollow struct {
-	Id          int       `json:"id" xorm:"autoincr pk not null"`
-	UserId      int       `json:"user_id" xorm:"not null"`
-	FollowId    int       `json:"follow_id" xorm:"not null"`
+	Id          int64     `json:"id" xorm:"autoincr pk not null"`
+	UserId      int64     `json:"user_id" xorm:"not null"`
+	FollowId    int64     `json:"follow_id" xorm:"not null"`
 	CreatedTime time.Time `json:"created_time" xorm:"created"`
 }
